"Of" qualifier now used for "Position Held" in auto description.
Previously the "Of" qualifier (P:642) was ignored when generating the
autodescription sentence for the "Position Held" property (P:39).  The
contents of the "Of" qualifier are now displayed at the end of the
sentence.  Ideally these would be positioned before the day and use the
term "of" but I wasn't able to work out the proper way to do that.

Bug: #8